Remove distinction between Widget and PVWidget.
