+++ src/ui/widgets/Display/display.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 
-import { Component, Widget, WidgetPropType } from "../widget";
+import { Widget } from "../widget";
+import { Component, WidgetPropType } from "../widgetProps";
 import { registerWidget } from "../register";
 import { ChoicePropOpt, ChildrenPropOpt, InferWidgetProps } from "../propTypes";
